move validation methods into trait
